app/utils/discord-webhook-server/backend/requirements.txt
discord.py>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
